backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Gmail API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Email processing
email-validator==2.1.0
beautifulsoup4==4.12.2
lxml==4.9.3

# AI/ML models (following PyTorch 2.8 requirement)
torch==2.8.0
torchvision
torchaudio
transformers==4.35.2
scikit-learn==1.3.2
spacy==3.7.2
sentence-transformers==2.2.2
huggingface-hub==0.16.4

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
pytz==2023.3

# Background tasks
celery==5.3.4
redis==5.0.1

# Search
elasticsearch==8.11.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
